集中タイマー値オブジェクトの実装#1

+++ src/main/java/com/example/domain/value_object/timer/TimerId.java
@@ -4,13 +4,13 @@ package com.example.domain.value_object.timer;
  * 集中タイマー
  */
 public class TimerId {
-    private Long timerId;
+    private int timerId;
 
-    public TimerId(Long timerId) {
+    public TimerId(int timerId) {
         this.timerId = timerId;
     }
 
-    public Long getTimerID() {
+    public int getTimerID() {
         return timerId;
     }
 }
